Fix all PHPStan (larastan) level 5 errors
Resolve the 193 static-analysis errors reported by larastan at level 5:

- Add generic PHPDoc `@return` types and native return type hints to all
  Eloquent relations so relation existence checks pass and related models
  keep their concrete type when iterated/accessed.
- Add `@property` annotations for plain database columns and transient
  attributes accessed dynamically on models (accessor-based attributes are
  detected automatically by larastan).
- Fix factory `$model` PHPDoc to `class-string<Model>` for covariance and
  add the missing `UnitFactory`.
- Correct `$fillable`/`$hidden`/`$casts` and provider property PHPDoc types.
- Cast `Cart::tax()`/`Cart::discount()` (string) to float before arithmetic.
- Access cart item options via array syntax to avoid the array|CartItemOptions
  union false positive.
- Replace uninstalled Intervention Image usage with a direct file store.
- Use `config()` instead of `env()` outside the config directory.
- Remove a redundant nested condition, an unused property, a double `DB::raw`
  wrap, an unnecessary nullsafe operator, and fix `str_pad` and flash() usages.

+++ Modules/PurchasesReturn/Http/Controllers/PurchasesReturnController.php
@@ -60,8 +60,8 @@ class PurchasesReturnController extends Controller
                 'payment_status' => $payment_status,
                 'payment_method' => $request->payment_method,
                 'note' => $request->note,
-                'tax_amount' => Cart::instance('purchase_return')->tax() * 100,
-                'discount_amount' => Cart::instance('purchase_return')->discount() * 100,
+                'tax_amount' => (float) Cart::instance('purchase_return')->tax() * 100,
+                'discount_amount' => (float) Cart::instance('purchase_return')->discount() * 100,
             ]);
 
             foreach (Cart::instance('purchase_return')->content() as $cart_item) {
@@ -184,8 +184,8 @@ class PurchasesReturnController extends Controller
                 'payment_status' => $payment_status,
                 'payment_method' => $request->payment_method,
                 'note' => $request->note,
-                'tax_amount' => Cart::instance('purchase_return')->tax() * 100,
-                'discount_amount' => Cart::instance('purchase_return')->discount() * 100,
+                'tax_amount' => (float) Cart::instance('purchase_return')->tax() * 100,
+                'discount_amount' => (float) Cart::instance('purchase_return')->discount() * 100,
             ]);
 
             foreach (Cart::instance('purchase_return')->content() as $cart_item) {
