+++ Modules/SalesReturn/Http/Controllers/SalesReturnController.php
@@ -60,8 +60,8 @@ class SalesReturnController extends Controller
                 'payment_status' => $payment_status,
                 'payment_method' => $request->payment_method,
                 'note' => $request->note,
-                'tax_amount' => Cart::instance('sale_return')->tax() * 100,
-                'discount_amount' => Cart::instance('sale_return')->discount() * 100,
+                'tax_amount' => (float) Cart::instance('sale_return')->tax() * 100,
+                'discount_amount' => (float) Cart::instance('sale_return')->discount() * 100,
             ]);
 
             foreach (Cart::instance('sale_return')->content() as $cart_item) {
@@ -184,8 +184,8 @@ class SalesReturnController extends Controller
                 'payment_status' => $payment_status,
                 'payment_method' => $request->payment_method,
                 'note' => $request->note,
-                'tax_amount' => Cart::instance('sale_return')->tax() * 100,
-                'discount_amount' => Cart::instance('sale_return')->discount() * 100,
+                'tax_amount' => (float) Cart::instance('sale_return')->tax() * 100,
+                'discount_amount' => (float) Cart::instance('sale_return')->discount() * 100,
             ]);
 
             foreach (Cart::instance('sale_return')->content() as $cart_item) {
